+++ sentry-kotlin-multiplatform/src/appleMain/kotlin/io/sentry/kotlin/multiplatform/nsexception/Throwable.kt
@@ -77,7 +77,8 @@ internal fun List<Long>.dropCommonAddresses(
 ): List<Long> {
     var i = commonAddresses.size
     if (i == 0) return this
+
     return dropLastWhile {
-        i-- >= 0 && commonAddresses[i] == it
+        i-- > 0 && commonAddresses[i] == it
     }
 }